requirements.txt
django==3.2
django-environ==0.5.0
psycopg2-binary==2.9.1
djangorestframework==3.12.4
django-nested-admin==3.3.3